change exec db to db.rs
