+++ crates/l2/prover/zkvm/interface/src/lib.rs
@@ -24,7 +24,7 @@ pub mod io {
         types::{Block, BlockHeader},
         H256,
     };
-    use ethrex_vm::backends::exec_db::ExecutionDB;
+    use ethrex_vm::db::ExecutionDB;
     use serde::{de::DeserializeOwned, Deserialize, Serialize};
     use serde_with::{serde_as, DeserializeAs, SerializeAs};
 
@@ -98,7 +98,7 @@ pub mod trie {
     use ethrex_rlp::{decode::RLPDecode, encode::RLPEncode, error::RLPDecodeError};
     use ethrex_storage::{hash_address, hash_key, AccountUpdate};
     use ethrex_trie::{Trie, TrieError};
-    use ethrex_vm::backends::exec_db::ExecutionDB;
+    use ethrex_vm::db::ExecutionDB;
     use thiserror::Error;
 
     #[derive(Debug, Error)]
