+++ crates/l2/sequencer/prover_server.rs
@@ -17,8 +17,8 @@ use ethrex_common::{
 use ethrex_l2_sdk::calldata::{encode_calldata, Value};
 use ethrex_rpc::clients::eth::{eth_sender::Overrides, EthClient, WrappedTransaction};
 use ethrex_storage::Store;
-use ethrex_vm::backends::exec_db::ExecutionDB;
 use ethrex_vm::backends::revm::execution_db::ToExecDB;
+use ethrex_vm::db::ExecutionDB;
 use ethrex_vm::{db::StoreWrapper, EvmError};
 use secp256k1::SecretKey;
 use serde::{Deserialize, Serialize};
