+++ crates/vm/backends/mod.rs
@@ -1,4 +1,3 @@
-pub mod exec_db;
 pub mod levm;
 pub mod revm;
